Package tlog client.
This packages the tlog client into a `tlog` package. The tlog client was
previously defined in the `tstore` package and was not exported.

You may notice that the newly added tlog package returns types from the
external trillian packages. Normally, I prefer to return only locally
defined types so that there is never any issues with upstream changes
impacting consumers, but in this case, it's not worth it to add the
locally defined types. Doing so would result in substantial code churn
in the politeiad codebase and Trillian is going to be replaced at some
point in the future anyway with a light weight, append-only merkle tree
that does not require running separate processes.
